fullstack-orthanc-dicom/order-management/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
requests==2.31.0
PyJWT==2.8.0
gunicorn==22.0.0
